+++ docs/components/DocsMenu.js
@@ -1,8 +1,14 @@
+import { Css } from "../../fachwerk.js";
+
 export default {
+  mixins: [Css],
   props: ["items", "value"],
-  data: () => ({ currentActiveItem: 0 }),
+  data: () => ({ currentActiveItem: 0, show: true }),
+  mounted() {
+    this.$global.$on("type", type => (this.show = type == "document"));
+  },
   template: `
-    <div>
+    <div v-show="show">
       <div
         v-for="(item,i) in items"
         :key="i"
@@ -13,10 +19,9 @@ export default {
         <div :style="{
           display: 'flex',
           alignItems: 'center',
-          padding: 'var(--base2) var(--base2) var(--base) var(--base6)',
+          padding: 'var(--base3) var(--base3) var(--base2) var(--base3)',
           fontWeight: 'bold',
           color: 'var(--primary)',
-          transform: 'translate(0,calc(var(--base) * 0))',
         }"
         @click="currentActiveItem = i;"
         v-html="item.title"
@@ -28,14 +33,27 @@ export default {
           :style="{
             display: 'flex',
             alignItems: 'center',
-            padding: 'var(--base) var(--base) var(--base) var(--base7)',
+            padding: 'var(--base) var(--base) var(--base) var(--base3)',
             border: 'none',
             fontWeight: 'normal'
           }"
+          class="router-link"
           :to="item.path"
-          v-html="item.title"
-        /> 
+        >
+          <span v-html="item.title" />
+        </router-link>
       </div>
+      <p />
+      <f-footer style="display: block;"/>
     </div>
+  `,
+  css: `
+    .router-link span {
+      border-bottom: 2px solid var(--transparent);
+    }
+    .router-link-exact-active span {
+      color: var(--blue);
+      border-bottom: 2px solid var(--blue);
+    }
   `
 };