+++ fachwerk.js
@@ -2,6 +2,7 @@ export * from './vendor.js'
 export * from './src/mixins.js'
 export * from './src/utils.js'
 export * from './src/fachwerk.js'
+export * from './src/fachwerk2.js'
 
 import * as components from './src/components.js'
 import * as utils from './src/utils.js'