+++ src/components/2d/FScene.js
@@ -12,12 +12,12 @@ export default {
   `,
   props: {
     width: {
-      default: 250,
+      default: 300,
       type: [Number, String],
       description: "Scene width in pixels"
     },
     height: {
-      default: 250,
+      default: 300,
       type: [Number, String],
       description: "Scene height in pixels"
     },